Fold toggleFighterSubtype into allowedFighterSubtypes
The new utils/fighter-subtypes.ts was too thin to justify its own module, and
its name sat awkwardly next to the three subtype files already in utils/.
allowedFighterSubtypes.ts covers the same domain — which subtypes a form may
offer — and is imported by the same forms, so the toggle helper belongs there.

Keeping it shared rather than inlining it: three components across two
directories call it, and the ordering invariant it preserves is the part that
is easy to reimplement wrongly.
